requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
xgboost>=1.5.0
lightgbm>=3.3.0
ta>=0.10.0
yfinance>=0.1.70
MetaTrader5>=5.0.0

# Feature engineering
pandas-ta==0.3.14b0

# Monitoring and visualization
plotly>=5.3.0
loguru>=0.5.3
mlflow>=1.20.2
matplotlib>=3.4.3
seaborn>=0.11.2
pytz>=2021.3

# Sampling and balancing
imbalanced-learn>=0.10.0

# Servo and execution dependencies
watchdog>=2.1.0
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
pyyaml>=6.0.0
joblib>=1.1.0
python-dotenv>=0.19.0
requests>=2.26.0

# Reinforcement Learning dependencies
stable-baselines3[extra]>=2.0.0
gym>=0.21.0
tensorboard>=2.8.0
torch>=1.12.0 